CodecLZOP

Documentation for CodecLZOP.

CodecLZOP.compress_block

CodecLZOP.decompress_block

CodecLZOP.compress_block — Method

compress_block(input, output, algo; [kwargs...])::Tuple{Int,Int}

Compress a block of data from `input` to `output` using LZO algorithm `algo`, returning the number of bytes read from `input` and written to `output`.

Arguments

input: An AbstractVector{UInt8} or IO object containing the block of data to compress.

output::IO: Output IO object to write the compressed block.

Keyword arguments

block_size::Integer = LZOP_DEFAULT_BLOCK_SIZE: Number of bytes to read from input. Will cap at LZOP_MAX_BLOCK_SIZE.

crc32::Bool = false: If true, write a CRC-32 checksum for both uncompressed and compressed data. If false, write Adler32 checksums instead.

filter_function::Function = identity: Transform the input data using the specified filter functions. The filter function must accept a single AbstractVector{UInt8} argument and must modify that vector in-place.

optimize::Bool = false: If true, process the data twice to optimize how it is stored for faster decompression. Setting this to true doubles compression time with little to no improvement in decompression time, so its use is not recommended.

CodecLZOP.decompress_block — Method

decompress_block(input, output, algo; [kwargs...])::Tuple{Int,Int}

Decompress a block of data from `input` to `output` using LZO algorithm `algo`, returning the number of bytes read from `input` and written to `output`.

Arguments

input: An AbstractVector{UInt8} or IO object containing the block of LZO-compressed data to decompress.

output::IO: Output IO object to write the decompressed block.

Keyword arguments

'crc32::Bool = false: Iftrue, assume the checksum written to the block for both uncompressed and compressed data is a CRC-32 checksum. Iffalse`, assume Adler32 checksums instead.

filter_function::Function = identity: Untransform the output data using the specified filter function. The filter function must take a single AbstractVector{UInt8} argument and modify it in place.

on_checksum_fail::Symbol = :throw: Choose how the function responds to invalud checksums. If :throw, an ErrorException will be thrown. If :warn, a warning will be printed. If :ignore, the checksum values will be completely ignored.